Big cleanup of UltimaGame
Moving a lot of framework code to CoreGame.cs
